feat(epreuves): searchable comboboxes for the resolve-dialog levels
Extract the concours combobox into a shared <SearchableSelect> and use it
for the épreuve établissement/filière/niveau/matière dropdowns too, so
every level filters as you type. Keeps the state-aware empty text
(Chargement… / Résolvez d'abord le niveau parent / Aucune … — créez-en une).
