+++ frontend/src/pages/ConcoursAdmin.tsx
@@ -29,59 +29,13 @@ import {
     DialogHeader,
     DialogTitle,
 } from "@/components/ui/dialog";
-import { Loader2, CheckCircle, XCircle, ChevronLeft, ChevronRight, Plus, Eye, Wrench, Check, AlertTriangle, ChevronsUpDown } from "lucide-react";
-import { Popover, PopoverContent, PopoverTrigger } from "@/components/ui/popover";
-import { Command, CommandEmpty, CommandGroup, CommandInput, CommandItem, CommandList } from "@/components/ui/command";
-import { cn } from "@/lib/utils";
+import { Loader2, CheckCircle, XCircle, ChevronLeft, ChevronRight, Plus, Eye, Wrench, Check, AlertTriangle } from "lucide-react";
+import { SearchableSelect } from "@/components/SearchableSelect";
 import { filesService } from "@/lib/services/files.service";
 import { concoursService, ConcoursSubmission } from "@/lib/services/concours.service";
 import { structureService } from "@/lib/services/structure.service";
 import { titreService } from "@/lib/services/titre.service";
 import { useToast } from "@/hooks/use-toast";
-
-// Searchable single-select (combobox) — structures/titres can be numerous, so
-// the plain <Select> is replaced by a filter-as-you-type list.
-function SearchableSelect({ value, options, onSelect, placeholder, searchPlaceholder, emptyText }: {
-    value?: number | null;
-    options: { id: number; nom: string }[];
-    onSelect: (id: number) => void;
-    placeholder: string;
-    searchPlaceholder: string;
-    emptyText: string;
-}) {
-    const [open, setOpen] = useState(false);
-    const [search, setSearch] = useState('');
-    const selected = options.find(o => o.id === value);
-    const filtered = search.trim()
-        ? options.filter(o => o.nom.toLowerCase().includes(search.trim().toLowerCase()))
-        : options;
-    return (
-        <Popover open={open} onOpenChange={(o) => { setOpen(o); if (!o) setSearch(''); }}>
-            <PopoverTrigger asChild>
-                <Button variant="outline" role="combobox" aria-expanded={open} className="w-full justify-between font-normal">
-                    <span className={selected ? '' : 'text-muted-foreground'}>{selected ? selected.nom : placeholder}</span>
-                    <ChevronsUpDown className="ml-2 h-4 w-4 shrink-0 opacity-50" />
-                </Button>
-            </PopoverTrigger>
-            <PopoverContent className="w-[--radix-popover-trigger-width] p-0">
-                <Command shouldFilter={false}>
-                    <CommandInput placeholder={searchPlaceholder} value={search} onValueChange={setSearch} />
-                    <CommandList>
-                        <CommandEmpty>{options.length === 0 ? emptyText : 'Aucun résultat pour cette recherche.'}</CommandEmpty>
-                        <CommandGroup>
-                            {filtered.map(o => (
-                                <CommandItem key={o.id} value={o.nom} onSelect={() => { onSelect(o.id); setOpen(false); setSearch(''); }}>
-                                    <Check className={cn("mr-2 h-4 w-4", value === o.id ? "opacity-100" : "opacity-0")} />
-                                    {o.nom}
-                                </CommandItem>
-                            ))}
-                        </CommandGroup>
-                    </CommandList>
-                </Command>
-            </PopoverContent>
-        </Popover>
-    );
-}
 
 // "Parameter" dialog (mirrors the épreuves ResolveDialog): resolve the
 // structure/titre (pick existing OR create) and edit année/lieu on a pending
